Only execute updatebot when os is linux

+++ cogs/update.py
@@ -4,6 +4,7 @@ from git import Repo
 import os
 import shutil
 import sys
+import platform
 import subprocess
 from stat import S_IREAD, S_IRGRP, S_IROTH, S_IWUSR
 from shutil import copyfile
@@ -16,32 +17,37 @@ class Update(commands.Cog):
     @commands.command()
     async def updatebot(self, ctx):
         """Attempts to update the bot directly from the GitHub repository."""
-        if str(ctx.message.author.id) == ownerID:
-            username = os.getlogin()
-            os.mkdir('/tmp/freeupdate')
-            HTTPS_REMOTE_URL = github_login_url
-            DEST_NAME = '/tmp/freeupdate'
-            cloned_repo = Repo.clone_from(HTTPS_REMOTE_URL, DEST_NAME)
-            dir_path = os.getcwd()
-            shutil.rmtree(dir_path + "/cogs/")
-            path = '/home/' + username
-            src = '/tmp/freeupdate/cogs'
-            dest = dir_path + "/cogs"
-            destination = shutil.copytree(src, dest)
-            copyfile('/tmp/freeupdate/bot.py', dir_path + '/bot.py')
-            copyfile('/tmp/freeupdate/freesetup.py', dir_path + '/freesetup.py')
-            copyfile('/tmp/freeupdate/README.md', dir_path + '/README.md')
-            shutil.rmtree('/tmp/freeupdate')
-            print("Done! Restart the bot to apply the changes!")
-            em = discord.Embed(title = "Updated!", description = "FreeDiscord updated! No error reported. Check your console to confirm this.")
-            em.add_field(name = "Note", value = "The bot will now shut down. Restart it to apply the changes.")
-            
-            await ctx.send(embed = em)
-            
-            await ctx.bot.close()
-            
+        operatingsys = platform.system()
+        if operatingsys == "Linux":
+            if str(ctx.message.author.id) == ownerID:
+                username = os.getlogin()
+                os.mkdir('/tmp/freeupdate')
+                HTTPS_REMOTE_URL = github_login_url
+                DEST_NAME = '/tmp/freeupdate'
+                cloned_repo = Repo.clone_from(HTTPS_REMOTE_URL, DEST_NAME)
+                dir_path = os.getcwd()
+                shutil.rmtree(dir_path + "/cogs/")
+                path = '/home/' + username
+                src = '/tmp/freeupdate/cogs'
+                dest = dir_path + "/cogs"
+                destination = shutil.copytree(src, dest)
+                copyfile('/tmp/freeupdate/bot.py', dir_path + '/bot.py')
+                copyfile('/tmp/freeupdate/freesetup.py', dir_path + '/freesetup.py')
+                copyfile('/tmp/freeupdate/README.md', dir_path + '/README.md')
+                shutil.rmtree('/tmp/freeupdate')
+                print("Done! Restart the bot to apply the changes!")
+                em = discord.Embed(title = "Updated!", description = "FreeDiscord updated! No error reported. Check your console to confirm this.")
+                em.add_field(name = "Note", value = "The bot will now shut down. Restart it to apply the changes.")
+
+                await ctx.send(embed = em)
+
+                await ctx.bot.close()
+
+            else:
+                em = discord.Embed(title = "This command is for the bot owner only.")
+                await ctx.send(embed = em)
         else:
-            em = discord.Embed(title = "This command is for the bot owner only.")
+            em = discord.Embed(title = "`updatebot` isn't supported for your OS.")
             await ctx.send(embed = em)
 
 def setup(bot):
